dashboards/requirements.txt
# Core dependencies
streamlit>=1.31.0
duckdb>=0.9.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
prophet>=1.1.4

# Visualization
plotly>=5.16.1

# API and data access
requests==2.31.0
python-dotenv==1.0.0
